Buffer.slice → Buffer.subarray, move encoding judgement to utils
